pyproject.toml
[project]
name = "yt-story"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.56.0",
    "ollama>=0.6.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "tiktoken>=0.12.0",
    "youtube-transcript-api>=1.2.3",
]
